pyproject.toml
[project]
name = "Stegastatter"
version = "1.0.2"
description = "A steganography engine"
authors = [
    {name = "Jebbex1"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["numpy (>=2.3.1,<3.0.0)", "reedsolo (>=1.7.0,<2.0.0)", "pycryptodome (>=3.23.0,<4.0.0)", "pillow (>=11.3.0,<12.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"